requirements.txt
Flask==3.0.0
gunicorn==23.0.0
Flask-Mail==0.9.1
python-dotenv==1.0.1
requests==2.32.3
qrcode[pil]
python-dotenv
Pillow
bakong_khqr
